moved non-relational inequalities and negation to Logos.CoreNonRel and pointed deps to here
