+++ lib/logos.ex
@@ -5,7 +5,8 @@ defmodule Logos do
 
   defmacro __using__(_opts \\ []) do
     quote do
-      import Logos.Core, only: [all: 1, any: 1, equal: 2, failure: 0, negate: 1, success: 0]
+      import Logos.Core, only: [all: 1, any: 1, equal: 2, failure: 0, success: 0]
+      import Logos.CoreNonRel, only: [gt: 2, gte: 2, lt: 2, lte: 2, negate: 1]
       import Logos.Interface, only: [ask: 2, choice: 1, defrule: 2, fork: 1, neg: 1, with_vars: 2]
       import Logos.Rule
     end
