+++ test/core_non_rel_test.exs
@@ -1,0 +1,4 @@
+defmodule CoreNonRelTest do
+  use ExUnit.Case, async: true
+  doctest Logos.CoreNonRel
+end
